+++ confik/CHANGELOG.md
@@ -3,6 +3,13 @@
 ## Unreleased
 
 - Implement `Configuration` for [`ahash::{AHashSet, AHashMap}`](https://docs.rs/ahash/0.8.11/ahash/).
+- Add new `helper` module, with utilities for manually implementing more complex `Configuration` behaviour.
+  - `UnkeyedContainerBuilder` can be used as a `Configuration::builder` for container types without separate keys (such as a `Vec` and `HashSet`).
+    - See `UnkeyedContainerBuilder`'s docs for details and an example.
+  - `KeyedContainerBuilder` can be used as a `Configuration::builder` for container types with explicit keys (such as `HashMap` and `BTreeMap`).
+    - Using `KeyedContainerBuilder` requires implementing `KeyedContainer` for your type.
+    - See `KeyedContainerBuilder`'s docs for details and an example.
+  - A few type aliases, to make it easier to write and understand complex generics when manually implementing `Configuration`.
 
 ## 0.14.0